# openapi.yaml
openapi: 3.0.0
info:
  title: Mitosis API
  description: |
    Mitosis is a unified distributed transport evaluation framework that provides REST API endpoints 
    for managing users, groups, workers, and tasks in a distributed computing environment.

    ## Authentication
    Most endpoints require JWT bearer token authentication. Get a token by calling `/login` with 
    valid credentials.

    ## API Structure
    - `/auth` - User authentication status
    - `/login` - User login
    - `/users` - User management
    - `/groups` - Group management and file attachments
    - `/workers` - Worker registration and management
    - `/tasks` - Task submission and monitoring
    - `/admin` - Administrative operations (admin users only)
    - `/health` - Service health check
    - `/redis` - Redis connection information
  version: 0.6.0
  contact:
    name: Mitosis
    url: https://github.com/stack-rs/mitosis
  license:
    name: Apache 2.0
    url: https://github.com/stack-rs/mitosis/blob/main/LICENSE
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.mitosis.example.com
    description: Production server
paths:
...
  /workers/tasks:
    get:
      summary: Fetch task for worker
      description: Worker fetches next available task
      operationId: fetchTask
      tags:
        - Workers
      security:
        - workerAuth: []
      responses:
        "200":
          description: Task assigned to worker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerTaskResp"
        "204":
          description: No tasks available
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Report task status
      description: Worker reports task execution status
      operationId: reportTask
      tags:
        - Workers
      security:
        - workerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportTaskReq"
      responses:
        "200":
          description: Status reported successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportTaskResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /workers/tasks/{uuid}:
    get:
      summary: Get task details for worker
      description: Worker queries task details
      operationId: getTaskForWorker
      tags:
        - Workers
      security:
        - workerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerTaskResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /workers/tasks/{uuid}/artifacts/{content_type}:
    get:
      summary: Download task artifact for worker
      description: Worker downloads task artifacts
      operationId: downloadTaskArtifactForWorker
      tags:
        - Workers
      security:
        - workerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: content_type
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ArtifactContentType"
      responses:
        "302":
          description: Redirect to download URL
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /tasks:
    post:
      summary: Submit task
      description: Submit a new task for execution
      operationId: submitTask
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitTaskReq"
      responses:
        "200":
          description: Task submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitTaskResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /tasks/query:
    post:
      summary: Query tasks
      description: List tasks with optional filtering
      operationId: queryTasks
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TasksQueryReq"
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TasksQueryResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /tasks/cancel:
    post:
      summary: Batch cancel tasks
      description: Cancel multiple tasks by filter criteria
      operationId: cancelTasksByFilter
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TasksCancelByFilterReq"
      responses:
        "200":
          description: Tasks cancelled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TasksCancelByFilterResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /tasks/cancel/list:
    post:
      summary: Batch cancel tasks by UUIDs
      description: Cancel multiple tasks by specific UUIDs
      operationId: cancelTasksByUuids
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TasksCancelByUuidsReq"
      responses:
        "200":
          description: Tasks cancelled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TasksCancelByUuidsResp"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /tasks/download/artifacts:
    post:
      summary: Batch download artifacts by filter
      description: Get download URLs for task artifacts matching filter criteria
      operationId: batchDownloadArtifactsByFilter
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtifactsDownloadByFilterReq"
      responses:
        "200":
          description: List of artifact download URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactsDownloadListResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /tasks/download/artifacts/list:
    post:
      summary: Batch download artifacts by UUIDs
      description: Get download URLs for task artifacts by specific UUIDs
      operationId: batchDownloadArtifactsByUuids
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtifactsDownloadByUuidsReq"
      responses:
        "200":
          description: List of artifact download URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactsDownloadListResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /tasks/{uuid}:
    get:
      summary: Get task information
      description: Get detailed information about a specific task
      operationId: getTask
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskQueryResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update task
      description: Modify task configuration (only for pending tasks)
      operationId: updateTask
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeTaskReq"
      responses:
        "200":
          description: Task updated successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

    delete:
      summary: Cancel task
      description: Cancel a pending or running task
      operationId: cancelTask
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task cancelled successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /tasks/{uuid}/labels:
    put:
      summary: Update task labels
      description: Replace task labels
      operationId: updateTaskLabels
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskLabelsReq"
      responses:
        "200":
          description: Labels updated successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /tasks/{uuid}/artifacts:
    post:
      summary: Upload task artifact
      description: Get upload URL for task artifacts
      operationId: uploadTaskArtifact
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadArtifactReq"
      responses:
        "200":
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadArtifactResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /tasks/{uuid}/artifacts/{content_type}:
    delete:
      summary: Delete task artifact
      description: Delete a task artifact
      operationId: deleteTaskArtifact
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: content_type
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ArtifactContentType"
      responses:
        "200":
          description: Artifact deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /tasks/{uuid}/download/artifacts/{content_type}:
    get:
      summary: Download task artifact
      description: Get download URL for task artifacts
      operationId: downloadTaskArtifact
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: content_type
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ArtifactContentType"
      responses:
        "302":
          description: Redirect to download URL
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/tasks/{uuid}/artifacts/{content_type}:
    delete:
      summary: Delete task artifact (Admin)
      description: Delete any task artifact (admin only)
      operationId: adminDeleteTaskArtifact
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: content_type
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ArtifactContentType"
      responses:
        "200":
          description: Artifact deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /login endpoint
    workerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained when registering worker
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request format
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid or expired token
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Insufficient permissions
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found
    Conflict:
      description: Conflict - resource already exists or state conflict
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource already exists
  schemas:
...
    ArtifactContentType:
      type: string
      enum:
        - result
        - exec-log
        - std-log
      description: |
        Type of task artifact:
        - result: Task results from MITO_RESULT_DIR
        - exec-log: Execution logs from MITO_EXEC_DIR
        - std-log: Standard output and error logs
...
    SubmitTaskReq:
      type: object
      required:
        - group_name
        - tags
        - labels
        - timeout
        - priority
        - task_spec
      properties:
        group_name:
          type: string
          description: Target group for task execution
          example: research-team
        tags:
          type: array
          items:
            type: string
          description: Required worker tags for task execution
          example:
            - gpu
            - python
        labels:
          type: array
          items:
            type: string
          description: Labels for task organization and querying
          example:
            - experiment-1
            - training
        timeout:
          type: string
          description: Task timeout duration (e.g. "1h", "30m")
          example: 2h
        priority:
          type: integer
          format: int32
          description: Task priority (higher = more important)
          example: 10
        task_spec:
          $ref: "#/components/schemas/TaskSpec"
    SubmitTaskResp:
      type: object
      required:
        - task_id
        - uuid
      properties:
        task_id:
          type: integer
          format: int64
          description: Internal task ID
          example: 12345
        uuid:
          type: string
          format: uuid
          description: Unique task identifier
          example: 789e0123-e89b-12d3-a456-426614174222
...
    TasksQueryReq:
      type: object
      properties:
        creator_usernames:
          type: array
          items:
            type: string
          nullable: true
          description: Filter by task creators
          example:
            - johndoe
            - janedoe
        group_name:
          type: string
          nullable: true
          description: Filter by group name
          example: research-team
        tags:
          type: array
          items:
            type: string
          nullable: true
          description: Filter by task tags
          example:
            - gpu
            - training
        labels:
          type: array
          items:
            type: string
          nullable: true
          description: Filter by task labels
          example:
            - experiment-1
        states:
          type: array
          items:
            $ref: "#/components/schemas/TaskState"
          nullable: true
          description: Filter by task states
        exit_status:
          type: string
          nullable: true
          description: Filter by exit status (e.g. "0", "!=0")
          example: "0"
        priority:
          type: string
          nullable: true
          description: Filter by priority (e.g. ">5", "<=10")
          example: ">5"
        limit:
          type: integer
          format: uint64
          nullable: true
          description: Maximum number of results
          example: 50
        offset:
          type: integer
          format: uint64
          nullable: true
          description: Offset for pagination
          example: 0
        count:
          type: boolean
          description: Whether to return total count
          default: false
...
    TaskQueryResp:
      type: object
      required:
        - info
        - artifacts
      properties:
        info:
          $ref: "#/components/schemas/ParsedTaskQueryInfo"
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactQueryResp"
          description: Available artifacts for this task
    TasksQueryResp:
      type: object
      required:
        - count
        - tasks
        - group_name
      properties:
        count:
          type: integer
          format: uint64
          description: Total number of tasks (if count=true)
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskQueryInfo"
        group_name:
          type: string
...
    WorkerTaskResp:
      type: object
      required:
        - id
        - uuid
        - timeout
        - spec
      properties:
        id:
          type: integer
          format: int64
          description: Internal task ID
        uuid:
          type: string
          format: uuid
          description: Task UUID
        timeout:
          type: string
          description: Task timeout duration
          example: 2h
        spec:
          $ref: "#/components/schemas/TaskSpec"
    ReportTaskReq:
      type: object
      required:
        - id
        - op
      properties:
        id:
          type: integer
          format: int64
          description: Internal task ID
        op:
          $ref: "#/components/schemas/ReportTaskOp"
...
    ReportTaskResp:
      type: object
      properties:
        url:
          type: string
          format: uri
          nullable: true
          description: Upload URL (only for Upload operations)
...
    TasksCancelByFilterReq:
      type: object
      properties:
        creator_usernames:
          type: array
          items:
            type: string
          nullable: true
          description: Filter by task creators
        group_name:
          type: string
          nullable: true
          description: Filter by group name
        tags:
          type: array
          items:
            type: string
          nullable: true
          description: Filter by task tags
        labels:
          type: array
          items:
            type: string
          nullable: true
          description: Filter by task labels
        states:
          type: array
          items:
            $ref: "#/components/schemas/TaskState"
          nullable: true
          description: Filter by task states
        exit_status:
          type: string
          nullable: true
          description: Filter by exit status
        priority:
          type: string
          nullable: true
          description: Filter by priority
      description: Request to cancel multiple tasks by filter criteria
    TasksCancelByFilterResp:
      type: object
      required:
        - cancelled_count
        - group_name
      properties:
        cancelled_count:
          type: integer
          format: uint64
          description: Number of tasks cancelled
          example: 15
        group_name:
          type: string
          description: Group name where tasks were cancelled
          example: research-team
    TasksCancelByUuidsReq:
      type: object
      required:
        - uuids
      properties:
        uuids:
          type: array
          items:
            type: string
            format: uuid
          description: List of task UUIDs to cancel
          example:
            - "550e8400-e29b-41d4-a716-446655440000"
            - "550e8400-e29b-41d4-a716-446655440001"
      description: Request to cancel multiple tasks by UUIDs
    TasksCancelByUuidsResp:
      type: object
      required:
        - cancelled_count
        - failed_uuids
      properties:
        cancelled_count:
          type: integer
          format: uint64
          description: Number of tasks successfully cancelled
          example: 10
        failed_uuids:
          type: array
          items:
            type: string
            format: uuid
          description: List of task UUIDs that failed to cancel (not found or no permission)
          example:
            - "550e8400-e29b-41d4-a716-446655440005"
...
    ArtifactsDownloadByFilterReq:
      type: object
      required:
        - content_type
      properties:
        creator_usernames:
          type: array
          items:
            type: string
          nullable: true
          description: Filter by task creators
        group_name:
          type: string
          nullable: true
          description: Filter by group name
        tags:
          type: array
          items:
            type: string
          nullable: true
          description: Filter by task tags
        labels:
          type: array
          items:
            type: string
          nullable: true
          description: Filter by task labels
        states:
          type: array
          items:
            $ref: "#/components/schemas/TaskState"
          nullable: true
          description: Filter by task states
        exit_status:
          type: string
          nullable: true
          description: Filter by exit status
        priority:
          type: string
          nullable: true
          description: Filter by priority
        content_type:
          $ref: "#/components/schemas/ArtifactContentType"
      description: Request to batch download artifacts by filter criteria
    ArtifactsDownloadByUuidsReq:
      type: object
      required:
        - uuids
        - content_type
      properties:
        uuids:
          type: array
          items:
            type: string
            format: uuid
          description: List of task UUIDs to download artifacts from
          example:
            - 789e0123-e89b-12d3-a456-426614174222
            - 123e4567-e89b-12d3-a456-426614174000
        content_type:
          $ref: "#/components/schemas/ArtifactContentType"
      description: Request to batch download artifacts by task UUIDs
...
    ArtifactsDownloadListResp:
      type: object
      required:
        - downloads
      properties:
        downloads:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactDownloadItem"
          description: List of artifact downloads
...
security:
  - bearerAuth: []
...